+++ CHANGELOG.rst
@@ -1,12 +1,16 @@
 Changelog
 =========
 
-master
-------
+v0.14.0
+-------
 
 - (`#190 <https://github.com/openscm/scmdata/pull/190>`_) Add special case for extrapolating timeseries containing a single timestep using ``constant`` extrapolation. Moved :attr:`scmdata.errors.InsufficientDataError` from :mod:`scmdata.time` to :mod:`scmdata.errors`
 - (`#186 <https://github.com/openscm/scmdata/pull/186>`_ and `#187 <https://github.com/openscm/scmdata/pull/187>`_) Fix the handling of non-alphanumeric characters in filenames on Windows for :class:`scmdata.database.ScmDatabase`. ``*`` values are no longer included in :class:`scmdata.database.ScmDatabase` filenames
 - (`#186 <https://github.com/openscm/scmdata/pull/186>`_ Move to ``pyproject.toml`` for setup etc.
+
+v0.13.2
+-------
+
 - (`#185 <https://github.com/openscm/scmdata/pull/185>`_) Allow :class:`scmdata.run.ScmRun` to read remote files by providing a URL to the constructor
 - (`#183 <https://github.com/openscm/scmdata/pull/183>`_) Deprecate :func:`scmdata.ops.integrate`, replacing with to :func:`scmdata.ops.cumsum` and :func:`scmdata.ops.cumtrapz`
 - (`#184 <https://github.com/openscm/scmdata/pull/184>`_) Add :func:`scmdata.run.ScmRun.round`